separated login screen from main
